Fix pathless route's match when parent is null  (#5964)
* Fix pathless route's match when parent is null

The logic for this has been moved to matchPath. When automatic path resolving is added, matchPath will need access to the parent match, so we will need to pass that argument to matchPath eventually anyways.

When a parent match is null, a match object with default values is returned by matchPath. This match object should not be relied upon for resolving paths, but it will allow a <Route> that uses the component or render prop to render.

Fixes #4695

* Make the default match obvious

* Add some match docs about null match

* Pathless child routes inherit parent, even when null

This is only an issue when you use <Route children>. If you render a pathless <Route> inside of a <Route children>, you must also use <Route children> (instead of <Route render> or <Route component>).

This commit also switches withRouter to using <Route children> instead of <Route render> so that it works when the parent match is null.

* Clarify docs

* Fix prop-types usage and format with Prettier